backend/requirements.txt
Django>=5,<6
djangorestframework>=3.14,<4
djangorestframework-simplejwt>=5.3,<6
django-cors-headers>=4.3,<5
mysqlclient>=2.2,<3
bcrypt>=4,<5
